OOP Online practice

+++ CSE/Online & Offline/1.1/B_online_2.c
@@ -2,11 +2,11 @@
 #include <stdlib.h>
 #include <stddef.h>
 
-int *merge_unique(const int *a, size_t n, const int *b, size_t m, size_t *out_len)
+int *merge_unique(const int *a, int n, const int *b, int m, int *out_len)
 {
-    size_t max_size = n + m;
+    int max_size = n + m;
     int *result = (int *)malloc(max_size * sizeof(int));
-    size_t i = 0, j = 0, k = 0;
+    int i = 0, j = 0, k = 0;
 
     while (i < n && j < m)
     {
@@ -56,26 +56,26 @@ int *merge_unique(const int *a, size_t n, const int *b, size_t m, size_t *out_le
 
 int main()
 {
-    size_t n, m;
+    int n, m;
     scanf("%zu", &n);
     int *a = (int *)malloc(n * sizeof(int));
-    for (size_t i = 0; i < n; i++)
+    for (int i = 0; i < n; i++)
     {
         scanf("%d", a + i);
     }
 
     scanf("%zu", &m);
     int *b = (int *)malloc(m * sizeof(int));
-    for (size_t i = 0; i < m; i++)
+    for (int i = 0; i < m; i++)
     {
         scanf("%d", b + i);
     }
 
-    size_t out_len;
+    int out_len;
     int *result = merge_unique(a, n, b, m, &out_len);
 
     printf("%zu\n", out_len);
-    for (size_t i = 0; i < out_len; i++)
+    for (int i = 0; i < out_len; i++)
     {
         printf("%d ", *(result + i));
     }
